GUI_detect/requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
Cython==0.29.21
easydict==1.9
Flask==1.1.2
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
NetSDK @ file:///C:/Users/happy/Desktop/dahuaSDK/General_NetSDK_ChnEng_Python_win64_IS_V3.052.0000002.0.R.201103/dist/NetSDK-2.0.0.1-py3-none-win_amd64.whl
numpy==1.18.5
opencv-python==4.4.0.46
pandas==1.2.1
Pillow==8.0.1
pycocotools==2.0.2
pygame==1.9.6
pyparsing==2.4.7
PyQt5==5.15.4
pyqt5-plugins==5.15.2.2.0.1
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pyqt5-tools==5.15.2.3.0.2
PyQtChart==5.15.4
PyQtChart-Qt5==5.15.2
pyqtgraph==0.11.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
PyYAML==5.4.1
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
requests==2.25.1
scipy==1.6.0
seaborn==0.11.1
six==1.15.0
thop==0.0.31.post2005241907
torch==1.6.0+cu101
torchvision==0.7.0+cu101
tornado==5.1.1
tqdm==4.56.0
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
wincertstore==0.2
